YH-2079: refactor
