+++ src/widgets/Specialization/KeywordsSection/ui/KeywordsSection/KeywordsSection.tsx
@@ -3,8 +3,8 @@ import { useTranslations } from 'next-intl';
 import { HhTopBySpecResponse } from '@/entities/hh';
 import { Specializations, i18Namespace } from '@/shared/config';
 import { Chip } from '@/shared/ui/Chip';
+import { SectionWrapper } from '@/shared/ui/SectionWrapper';
 import { CollapsibleChipGrid } from '@/widgets/Specialization/CollapsibleChipGrid';
-import { SectionWrapper } from '@/widgets/Specialization/SectionWrapper';
 
 interface KeywordsSectionProps {
 	keywords?: HhTopBySpecResponse['keywords'];
